bee2bee-indexer/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bee2bee-indexer"
version = "0.1.0"
description = "High-performance code indexing library for AI agents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "RepoMind Team"}
]
keywords = ["code-indexing", "semantic-search", "tree-sitter", "embeddings", "vector-database"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "httpx>=0.27.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",

    # Tree-sitter parsers
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tree-sitter-typescript>=0.21.0",
    "tree-sitter-rust>=0.21.0",
    "tree-sitter-go>=0.21.0",
    "tree-sitter-java>=0.21.0",
    "tree-sitter-cpp>=0.22.0",

    # Vector DB
    "chromadb>=0.4.22",

    # Embeddings
    "sentence-transformers>=2.3.0",
    "openai>=1.12.0",

    # GitHub integration
    "PyGithub>=2.1.1",

    # Utils
    "pydantic>=2.5.0",
    "inflection>=0.5.1",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/heronlancellot/hackglobal-2025"
Documentation = "https://github.com/heronlancellot/hackglobal-2025/tree/main/bee2bee-indexer"
Repository = "https://github.com/heronlancellot/hackglobal-2025"

[tool.setuptools]
packages = ["bee2bee_indexer"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
